add userId

+++ src/api/controllers/private/todos/listTodos.ts
@@ -1,8 +1,12 @@
 import { Context } from "koa";
+import { ObjectId } from "mongodb";
 import todosCollection from "../../../models/todosModels";
 
 export const getTodos = async (ctx: Context) => {
-  const todos = await todosCollection.find();
+  const { userId } = ctx.request.body;
+
+  const todos = await todosCollection.find({ userId: ObjectId(userId) });
+
   ctx.response.status = 200;
   ctx.body = {
     message: "Ok",
@@ -11,9 +15,12 @@ export const getTodos = async (ctx: Context) => {
 };
 
 export const checkAllTodos = async (ctx: Context) => {
-  const { check } = ctx.request.body;
-  await todosCollection.updateMany({ check });
+  const { check, userId } = ctx.request.body;
+
+  await todosCollection.updateMany({ check, userId: ObjectId(userId) });
+
   const todos = await todosCollection.find();
+
   ctx.response.status = 200;
   ctx.body = {
     message: "Ok",
@@ -22,8 +29,12 @@ export const checkAllTodos = async (ctx: Context) => {
 };
 
 export const deleteCompletedTodos = async (ctx: Context) => {
-  await todosCollection.deleteMany({ check: true });
+  const { userId } = ctx.request.body;
+
+  await todosCollection.deleteMany({ check: true, userId: ObjectId(userId) });
+
   const todos = await todosCollection.find();
+
   ctx.response.status = 200;
   ctx.body = {
     message: "Ok",
